Separate mod lifecycle events from mod entrypoint
This is to make each class more single-responsibility and disconnect any sided classes from unsided classes
