+++ src/main/java/de/guntram/mcmod/antighost/AntiGhost.java
@@ -1,36 +1,18 @@
 package de.guntram.mcmod.antighost;
 
-import com.mojang.blaze3d.platform.InputConstants;
-import net.minecraft.client.KeyMapping;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.client.event.RegisterKeyMappingsEvent;
-import net.minecraftforge.client.settings.KeyConflictContext;
-import net.minecraftforge.common.util.Lazy;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import org.lwjgl.glfw.GLFW;
+import net.minecraftforge.fml.loading.FMLEnvironment;
 
-// Subscribes to mod lifecycle events
-@Mod(AntiGhost.MODID)
-@Mod.EventBusSubscriber(modid = AntiGhost.MODID, bus = Mod.EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
+import static de.guntram.mcmod.antighost.AntiGhost.MOD_ID;
+
+@Mod(MOD_ID)
 public class AntiGhost {
-    public static final String MODID = "antighost";
+    public static final String MOD_ID = "antighost";
+    public static final String MOD_NAME = "AntiGhost";
 
-    // Lazy initialize the key mapping to refresh blocks
-    // Only conflicts in-game
-    public static final Lazy<KeyMapping> REFRESH_BLOCKS_MAPPING = Lazy.of(
-            () -> new KeyMapping(
-                    "key." + MODID + ".reveal",
-                    KeyConflictContext.IN_GAME,
-                    InputConstants.Type.KEYSYM,
-                    GLFW.GLFW_KEY_G,
-                    "key.categories." + MODID
-            )
-    );
-
-    // Register the key binding for refreshing
-    @SubscribeEvent
-    public static void registerBindings(RegisterKeyMappingsEvent e) {
-        e.register(REFRESH_BLOCKS_MAPPING.get());
+    public AntiGhost() {
+        if (FMLEnvironment.dist.isDedicatedServer()) {
+            System.err.println(MOD_NAME + " detected a dedicated server. Not doing anything.");
+        }
     }
 }
